requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
openai==1.8.0
